+++ src/node/desktop/src/main/main-window.ts
@@ -13,14 +13,14 @@
  *
  */
 
-import { BrowserWindow, session } from 'electron';
+import { BrowserWindow, Menu, session } from 'electron';
 import path from 'path';
 import { ChildProcess } from 'child_process';
 
 import { logger } from '../core/logger';
 
 import { GwtCallback, PendingQuit } from './gwt-callback';
-import { MenuCallback } from './menu-callback';
+import { MenuCallback, showPlaceholderMenu } from './menu-callback';
 import { PendingWindow } from './pending-window';
 import { RCommandEvaluator } from './r-command-evaluator';
 import { SessionLauncher } from './session-launcher';
@@ -36,13 +36,83 @@ export class MainWindow extends GwtWindow {
   quitConfirmed = false;
   workbenchInitialized = false;
   pendingWindows = new Array<PendingWindow>();
+  private isErrorDisplayed = false;
+
+  // TODO
+  //#ifdef _WIN32
+  // HWINEVENTHOOK eventHook_ = nullptr;
+  //#endif
 
   constructor(url: string, public isRemoteDesktop: boolean) {
     super(false, false, '', url, undefined, undefined, isRemoteDesktop, ['desktop', 'desktopMenuCallback']);
+
     appState().gwtCallback = new GwtCallback(this, isRemoteDesktop);
-    this.menuCallback = new MenuCallback(this);
+    this.menuCallback = new MenuCallback();
 
     RCommandEvaluator.setMainWindow(this);
+
+    if (this.isRemoteDesktop) {
+      // TODO - determine if we need to replicate this
+      // since the object registration is asynchronous, during the GWT setup code
+      // there is a race condition where the initialization can happen before the
+      // remoteDesktop object is registered, making the GWT application think that
+      // it should use regular desktop objects - to circumvent this, we use a custom
+      // user agent string that the GWT code can detect with 100% success rate to
+      // get around this race condition
+      // QString userAgent = webPage()->profile()->httpUserAgent().append(QStringLiteral("; RStudio Remote Desktop"));
+      // webPage()->profile()->setHttpUserAgent(userAgent);
+      // channel->registerObject(QStringLiteral("remoteDesktop"), &gwtCallback_);
+    }
+
+    showPlaceholderMenu();
+
+    this.menuCallback.on(MenuCallback.MENUBAR_COMPLETED, (menu: Menu) => {
+      Menu.setApplicationMenu(menu);
+    });
+    this.menuCallback.on(MenuCallback.COMMAND_INVOKED, (commandId) => {
+      this.invokeCommand(commandId);
+    });
+
+    // TODO
+    // connect(&menuCallback_, SIGNAL(zoomActualSize()), this, SLOT(zoomActualSize()));
+    // connect(&menuCallback_, SIGNAL(zoomIn()), this, SLOT(zoomIn()));
+    // connect(&menuCallback_, SIGNAL(zoomOut()), this, SLOT(zoomOut()));
+    // connect(&gwtCallback_, SIGNAL(workbenchInitialized()),
+    //         this, SIGNAL(firstWorkbenchInitialized()));
+
+    appState().gwtCallback?.on(GwtCallback.WORKBENCH_INITIALIZED, () => {
+      this.onWorkbenchInitialized();
+    });
+
+    // connect(&gwtCallback_, SIGNAL(sessionQuit()),
+    //         this, SLOT(onSessionQuit()));
+
+    // connect(webView(), SIGNAL(onCloseWindowShortcut()),
+    //         this, SLOT(onCloseWindowShortcut()));
+
+    // connect(webView(), &WebView::urlChanged,
+    //         this, &MainWindow::onUrlChanged);
+   
+    this.webView.webContents.on('did-finish-load', () => {
+      this.onLoadFinished(true);
+    });
+    this.webView.webContents.on('did-fail-load', () => {
+      this.onLoadFinished(false);
+    });
+
+    // TODO
+    // connect(webPage(), &QWebEnginePage::loadFinished,
+    //         &menuCallback_, &MenuCallback::cleanUpActions);
+
+    // connect(&desktopInfo(), &DesktopInfo::fixedWidthFontListChanged, [this]() {
+    //    QString js = QStringLiteral(
+    //       "if (typeof window.onFontListReady === 'function') window.onFontListReady()");
+    //    this->webPage()->runJavaScript(js);
+    // });
+
+    // connect(qApp, SIGNAL(commitDataRequest(QSessionManager&)),
+    //         this, SLOT(commitDataRequest(QSessionManager&)),
+    //         Qt::DirectConnection);
   }
 
   loadUrl(url: string): void {
@@ -79,7 +149,20 @@ export class MainWindow extends GwtWindow {
   }
 
   invokeCommand(cmdId: string): void {
-    this.executeJavaScript(`window.desktopHooks.invokeCommand("${cmdId}")`)
+    let cmd = '';
+    if (process.platform === 'darwin') {
+      cmd = ` 
+        var wnd;
+        try {
+          wnd = window.$RStudio.last_focused_window;
+        } catch (e) {
+          wnd = window;
+        }
+        (wnd || window).desktopHooks.invokeCommand('${cmdId}');`;
+    } else {
+      cmd = `window.desktopHooks.invokeCommand("${cmdId}")`;
+    }
+    this.executeJavaScript(cmd)
       .catch((error) => {
         logger().logError(error);
       });
@@ -123,5 +206,18 @@ export class MainWindow extends GwtWindow {
 
   prepareForWindow(pendingWindow: PendingWindow): void {
     this.pendingWindows.push(pendingWindow);
+  }
+
+  onActivated(): void {
+    // intentionally left blank
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  onLoadFinished(ok: boolean): void {
+    // TODO
+  }
+
+  setErrorDisplayed(): void {
+    this.isErrorDisplayed = true;
   }
 }