+++ NEWS.md
@@ -37,6 +37,7 @@
 
 * **BREAKING:** RStudio Desktop Pro only supports activation with license files (Pro #2300)
 * Improved ordering of completion results within `library()` calls (#9293)
+* Added support for publishing Quarto documents and websites (#9556)
 * Add option to synchronize the Files pane with the current working directory in R (#4615)
 * Add new *Set Working Directory* command to context menu for source files (#6781)
 * Local background jobs can now be replayed (#5548)
@@ -50,3 +51,4 @@
 * Add file uploads and downloads to session audit log (Pro #2226)
 * RStudio no longer treats R objects containing null external pointers specially when building Environment pane (#5546)
 * Make Cmd+Shift+0 the shortcut for restarting session on MacOS (#7695)
+* RStudio Workbench can now run without root access in a single-user mode (Pro #9496)